+++ src/Subtitles/suphd.h
@@ -64,8 +64,8 @@ public:
     SubPicture *getSubPicture(int index);
 
 signals:
-    void maxProgressChanged(int maxProgress);
-    void currentProgressChanged(int currentProgress);
+    void maxProgressChanged(long maxProgress);
+    void currentProgressChanged(long currentProgress);
 
 private:
     int primaryColorIndex = 0;
